#36-어댑터/클래스 정리
