+++ app/src/main/java/com/teamnexters/lastwednesday/model/MyData.java
@@ -1,0 +1,15 @@
+package com.teamnexters.lastwednesday.model;
+
+/**
+ * Created by user on 2018-02-10.
+ */
+public class MyData{
+    public String text, text2, text3;
+    public int img;
+    public MyData(String text, String text2, String text3, int img){
+        this.text = text;
+        this.text2 = text2;
+        this.text3 = text3;
+        this.img=img;
+    }
+}
